extend mixin functionality to more layers
