+++ app/router.js
@@ -11,7 +11,7 @@ Router.map(function () { // eslint-disable-line
   this.route('lot', { path: 'lot/:boro/:block/:lot' });
   this.route('zma', { path: 'zma/:ulurpno' });
   this.route('bbl', { path: 'bbl/:bbl' });
-  this.route('zoning-district', { path: 'zoning-district/:zonedist' }); // make this zoning-district/:zonedist?search vs. ?click ???
+  this.route('zoning-district', { path: 'zoning-district/:zonedist' });
   this.route('special-purpose-district', { path: 'special-purpose-district/:id' });
   this.route('special-purpose-subdistricts', { path: 'special-purpose-subdistrict/:id' });
   this.route('commercial-overlay', { path: 'commercial-overlay/:id' });
