+++ app/routes/lot.js
@@ -1,9 +1,9 @@
 import Route from '@ember/routing/route';
 import { inject as service } from '@ember/service';
 import bblDemux from '../utils/bbl-demux';
-import updateSelectionLotMixin from '../mixins/update-selection-lot';
+import updateSelectionSingleFeatureMixin from '../mixins/update-selection-single-feature';
 
-export default Route.extend(updateSelectionLotMixin, {
+export default Route.extend(updateSelectionSingleFeatureMixin, {
   mainMap: service(),
 
   model(params) {
