+++ app/routes/special-purpose-district.js
@@ -1,9 +1,9 @@
 import Route from '@ember/routing/route';
-import { inject as service } from '@ember/service';
-import updateSelectionZoningMixin from '../mixins/update-selection-zoning';
+// import { inject as service } from '@ember/service';
+import updateSelectionAllFeaturesMixin from '../mixins/update-selection-all-features';
 
-export default Route.extend(updateSelectionZoningMixin, {
-  mainMap: service(),
+export default Route.extend(updateSelectionAllFeaturesMixin, {
+  // mainMap: service(),
   model(params) {
     return {
       taskInstance: this.store.findRecord('special-purpose-district', params.id),
